chore: add utility getImportEqualsDeclarations
