+++ src/transforms/v2-to-v3/modules/importEqualsModule/getImportEqualsDeclarations.ts
@@ -1,0 +1,25 @@
+import { Collection, JSCodeshift, StringLiteral, TSExternalModuleReference } from "jscodeshift";
+import { PACKAGE_NAME } from "../../config";
+
+export const getImportEqualsDeclarations = (
+  j: JSCodeshift,
+  source: Collection<unknown>,
+  path?: string
+) =>
+  source
+    .find(j.TSImportEqualsDeclaration, {
+      type: "TSImportEqualsDeclaration",
+      moduleReference: {
+        type: "TSExternalModuleReference",
+        expression: { type: "StringLiteral" },
+      },
+    })
+    .filter((importEqualsDeclaration) => {
+      const moduleReference = importEqualsDeclaration.value
+        .moduleReference as TSExternalModuleReference;
+      const expressionValue = (moduleReference.expression as StringLiteral).value;
+      if (path) {
+        return expressionValue === path;
+      }
+      return expressionValue.startsWith(PACKAGE_NAME);
+    });
