+++ src/transforms/v2-to-v3/modules/importEqualsModule/index.ts
@@ -1,2 +1,3 @@
 export * from "./addNamedModule";
 export * from "./getImportSpecifiers";
+export * from "./getImportEqualsDeclarations";
